feat(study): SJIP-1315 hide file section if no tables data
